Moved tests to the utils package and changed to CommonJS
