+++ .github/workflows/build.yml
@@ -15,11 +15,13 @@ jobs:
     - uses: actions/setup-node@v4
       with:
         node-version: 20
+    - run: npm --prefix users/utils ci
     - run: npm --prefix users/authservice ci
     - run: npm --prefix users/userservice ci
     - run: npm --prefix questionservice ci
     - run: npm --prefix gatewayservice ci
     - run: npm --prefix webapp ci
+    - run: npm --prefix users/utils test -- --coverage
     - run: npm --prefix users/authservice test -- --coverage
     - run: npm --prefix users/userservice test -- --coverage
     - run: npm --prefix questionservice test -- --coverage
