+++ users/authservice/test/auth-service.test.ts
@@ -4,8 +4,6 @@ import bcrypt from 'bcrypt';
 import User from '../src/models/auth-model';
 import mongoose from 'mongoose';
 import app from '../src/app';
-import { Request } from 'express';
-import {validateNotEmpty, validateRequiredLength} from "../../utils/src/field-validations";
 
 let mongoServer: MongoMemoryServer;
 
@@ -60,50 +58,4 @@ describe('Auth Service', () => {
     const response = await request(app).post('/login').send(invalidBody);
     expect(response.status).toBe(400);
   });
-
-  // ======================== Tests for utils ========================
-  // Note: this is duplicated code from the userservice, field-validations.ts
-  // should probably be placed outside and imported by both services, as well
-  // as the userSchema.
-
-  // Empty field validation
-  it('should get an error when passing an empty parameter', async () => {
-    const mockRequest = {
-      body: {}
-    } as Request;
-    mockRequest.body['history'] = '';
-
-    try {
-      validateNotEmpty(mockRequest, ['history']);
-      fail('Should get an error in the previous call');
-    } catch (error) {
-    }
-    // Should also get an error when the field does not exist
-    try {
-      validateNotEmpty(mockRequest, ['nonexistent']);
-      fail('Should get an error in the previous call');
-    } catch (error) {
-    }
-  });
-
-  // Empty field validation
-  it('should get an error when passing a parameter without the expected length', async () => {
-    const mockRequest = {
-      body: {}
-    } as Request;
-    mockRequest.body['test'] = '123456789';
-
-    try {
-      validateRequiredLength(mockRequest, ['test'], 10);
-      fail('Should get an error in the previous call');
-    } catch (error) {
-    }
-    // Should also get an error when the field does not exist
-    try {
-      validateRequiredLength(mockRequest, ['nonexistent'], 10);
-      fail('Should get an error in the previous call');
-    } catch (error) {
-    }
-  });
-  // =================================================================
 });
