+++ users/userservice/package.json
@@ -24,7 +24,7 @@
     "body-parser": "^1.20.2",
     "express": "^4.19.1",
     "jsonwebtoken": "^9.0.2",
-    "kaw-users-utils": "^1.0.0",
+    "kaw-users-utils": "^1.0.7",
     "mongoose": "^8.0.4"
   },
   "devDependencies": {
