+++ users/userservice/src/controllers/user-controller.ts
@@ -1,11 +1,11 @@
 import { Request, Response } from 'express';
 import bcrypt from 'bcrypt';
 
-import {
+const {
   validateRequiredFields,
   validateNotEmpty,
   validateRequiredLength,
-} from 'kaw-users-utils';
+} = require('kaw-users-utils');
 
 import User from '../models/user-model';
